# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.errorBody:
    properties:
      code:
        example: validation_error
        type: string
      message:
        example: Неверные данные запроса
        type: string
    type: object
  handler.errorResponse:
    properties:
      error:
        $ref: '#/definitions/handler.errorBody'
    type: object
  handler.eventDTO:
    properties:
      away:
        type: string
      home:
        type: string
      id:
        type: integer
      league_name:
        type: string
      markets:
        items:
          $ref: '#/definitions/handler.marketDTO'
        type: array
      source:
        type: string
      sport_slug:
        type: string
      starts_at:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  handler.eventsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.eventDTO'
        type: array
      page:
        type: integer
    type: object
...
  handler.leagueDTO:
    properties:
      id:
        type: integer
      name:
        type: string
      sport_slug:
        type: string
    type: object
  handler.leaguesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handler.leagueDTO'
        type: array
    type: object
...
  handler.marketDTO:
    properties:
      id:
        type: integer
      line:
        type: number
      outcomes:
        items:
          $ref: '#/definitions/handler.outcomeDTO'
        type: array
      question:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
  handler.outcomeDTO:
    properties:
      code:
        type: string
      id:
        type: integer
      label:
        type: string
      odds:
        type: number
      result:
        type: string
    type: object
...
  handler.sportsResponse:
    properties:
      sports:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: REST API платформы ставок на фантики. На текущем этапе (M1) реализованы
    аутентификация и профиль пользователя.
  title: FanticBet API
  version: "1.0"
paths:
...
  /events:
    get:
      description: Страница событий с рынками и текущими коэффициентами. Фильтры sport/status/league_id/q
        опциональны.
      parameters:
      - description: Фильтр по виду спорта (sport_slug)
        in: query
        name: sport
        type: string
      - description: Фильтр по источнику
        enum:
        - oddsapi
        - manual
        - custom
        in: query
        name: source
        type: string
      - description: Фильтр по статусу
        enum:
        - upcoming
        - live
        - settled
        - cancelled
        in: query
        name: status
        type: string
      - description: Фильтр по чемпионату (id)
        in: query
        name: league_id
        type: integer
      - description: Поиск по названию события
        in: query
        name: q
        type: string
      - default: 1
        description: Номер страницы (с 1)
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.eventsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Лента событий
      tags:
      - events
  /events/{id}:
    get:
      description: Событие со всеми рынками и исходами (текущие коэффициенты).
      parameters:
      - description: ID события
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.eventDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Событие
      tags:
      - events
...
  /leagues:
    get:
      description: Список чемпионатов (лиг) для фильтра ленты. Параметр sport_slug
        опционален.
      parameters:
      - description: Фильтр по виду спорта (sport_slug)
        in: query
        name: sport_slug
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.leaguesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Чемпионаты
      tags:
      - events
...
  /sports:
    get:
      description: Список видов спорта, по которым есть события в БД (всегда включает
        custom).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.sportsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Виды спорта
      tags:
      - events
...
securityDefinitions:
  BearerAuth:
    description: 'Введите access-токен в формате: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"